Fixed a duplicated Sid in policy Update terminator classes to match new policy layout
